apps/backend/cron/requirements.txt
yt-dlp
psycopg2-binary
python-dotenv
